requirements.txt
Flask
PyPDF2
firebase-admin
requests
python-dotenv
pinecone-client 